updated to be functional
There was trouble with some timing maybe because it is asynchronous and the loop is slow. I believe that it could be a problem but I am not sure. I tried stepping through but I would get stuck somewhere in the min.js. It may alos have something to do with the Live server? or I am failing to understand the problem in my code. It only catches the gameEnd() logic sometimes. and those checks are relatively inconsistent.